+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Skip GL draw function for EGL
 * Add an option to disable fastforward during input editor rewind (#491)
 * Add an option to keep rendering during fastforward (491)
+* Alert when an input column cannot be removed
 
 ### Changed
 ### Fixed